feat: add passport authentication and cookie handling, enhance auth controller and service

+++ src/tasks/tasks.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { Task } from '../schemas/tasks.schema';
 import { Model } from 'mongoose';
+import { Utils } from 'src/utils/Utils';
 
 @Injectable()
 export class TasksService {
@@ -18,14 +19,26 @@ export class TasksService {
   }
 
   findOne(id: string) {
+    if (!Utils.isObjectId(id)) {
+      throw new BadRequestException('Invalid ObjectId passed as id');
+    }
+
     return this.taskModel.findById(id);
   }
 
   update(id: string, updateTaskDto: UpdateTaskDto) {
+    if (!Utils.isObjectId(id)) {
+      throw new BadRequestException('Invalid ObjectId passed as id');
+    }
+
     return this.taskModel.findByIdAndUpdate(id, updateTaskDto);
   }
 
   remove(id: string) {
+    if (!Utils.isObjectId(id)) {
+      throw new BadRequestException('Invalid ObjectId passed as id');
+    }
+
     return this.taskModel.findByIdAndDelete(id);
   }
 }
